fix: Move commands into seperate directory and modules
