+++ src/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Discord command registration modules."""
